bousai_reminder/requirements.txt
requests
python-dotenv
schedule
